Fixes array cast incompatibility
For models with an array cast, the attributesToArray() method returns the cast array instead of the serialized array. So usage of `markApproved(), markRejected(), markPostponed() and markPending()` methods of the `Moderatable` trait throw a DB exception due to the attempt to save the raw array instead of the serialized array. So this fix changes the `$new->attributesToArray()` used in the mentioned methods to `$new->attributes`.